# docs/swagger.yaml
definitions:
  account.EmailPassword:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  account.Identity:
    properties:
      account:
        $ref: '#/definitions/account.ResponseAccount'
      profiles:
        items:
          $ref: '#/definitions/account.ResponseProfile'
        type: array
    type: object
  account.PublicResponse:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: object
    required:
    - email
    - id
    - password
    type: object
  account.ResponseAccount:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  account.ResponseProfile:
    properties:
      avatar:
        description: Avatar holds the value of the "avatar" field.
        type: string
      call:
        description: Call holds the value of the "call" field.
        type: string
      category:
        description: Category holds the value of the "category" field.
        type: string
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  account.ResponseToken:
    properties:
      body:
        description: Body holds the value of the "body" field.
        type: string
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      expire:
        description: Expire holds the value of the "expire" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      type:
        description: Type holds the value of the "type" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
...
  episode.Episode:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  episode.ResponseEpisode:
    properties:
      content:
        description: Content holds the value of the "content" field.
        type: string
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        type: object
      id:
        description: ID of the ent.
        type: string
      sid:
        description: Sid holds the value of the "sid" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  errs.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /account:
    get:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account.Identity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: info an account
      tags:
      - account
      - profile
  /account/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: account info
        in: body
        name: accountInfo
        required: true
        schema:
          $ref: '#/definitions/account.EmailPassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account.ResponseToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: login an account
      tags:
      - account
  /account/refresh:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account.ResponseToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: refresh token
      tags:
      - account
      - token
  /account/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: account info
        in: body
        name: accountInfo
        required: true
        schema:
          $ref: '#/definitions/account.EmailPassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/account.PublicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: register an account
      tags:
      - account
  /episode:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/episode.ResponseEpisode'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: get all episodes
      tags:
      - episode
  /episode/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: episode id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/episode.ResponseEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: get episode
      tags:
      - episode
  /episode/{pid}:
    post:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: episode info
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/episode.Episode'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/episode.ResponseEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: create episode
      tags:
      - episode
  /episode/{pid}/{id}:
    delete:
      parameters:
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: episode id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: delete one episode
      tags:
      - episode
    put:
      consumes:
      - application/json
      parameters:
      - description: episode id
        in: path
        name: id
        required: true
        type: string
      - description: profile id
        in: path
        name: pid
        required: true
        type: string
      - description: episode info
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/episode.Episode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/episode.ResponseEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: update episode
      tags:
      - episode
...
swagger: "2.0"